requirements.txt
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
loguru==0.7.3
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
urllib3==2.3.0
